board: evb_rk3588: add get rtc device address to select dtb for CM5
The cm5-io has the same hw_id value as the cm5-rpi-cm4-io,
and the corresponding dtb file needs to be selected by obtaining the address of the rtc device on the I2C bus.

The rtc device of cm5-io is on the I2C6 bus.
The rtc device for cm5-rpi-cm4-io is on the I2C7 bus.
Both rtc devices have address 0x51.

Signed-off-by: Mitchell Ma <machuang@radxa.com>